lab/requirements.txt
django
djangorestframework
Pillow
reportlab